requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.1.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
praw>=7.7.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0